feat: restructure utils and add LED configurations
- Move `Alliance` and `AutonomousRoutine` to a new `Autonomous` package
- Update imports in the main container to reflect package changes
- Add new `LEDConfiguration` enum for managing LED patterns
